Finished up security implementation.

+++ src/repository/mongo/entities/entity-repository-base.ts
@@ -1,9 +1,23 @@
 import { Entity } from '@common/models/entities/entity';
-import { IMongoEntityRepository } from './imongo-entity-repository';
+import {
+  IMongoEntityRepository,
+  IMongoRepository,
+} from './imongo-entity-repository';
 import { Db, Collection, FilterQuery } from 'mongodb';
 import shortid from 'shortid';
-import { IEntityRepositoryReferencePopulator } from './references/ientity-repository-reference-populator';
+import {
+  IEntityRepositoryReferencePopulator,
+  ReferenceType,
+  IEntityRepositorySingleReferencePopulator,
+  IEntityRepositoryManyReferencesPopulator,
+} from './references/ientity-repository-reference-populator';
 import { logger } from '@common/logging/winston';
+import { DbRequestContext } from '@app/database/db-request-context';
+import { PermissionsService } from '@common/security/permissions/permissions-service';
+import { CollectionPermission } from '@common/models/entities/permissions/collection-permission';
+import { PermissionAction } from '@common/security/permissions/permission';
+import { IPermissionWaiver } from '@common/security/permissions/ipermission-waiver';
+import { EmptyUpdateEntityError } from '@app/errors/repository/empty-update-entity-error';
 
 const noMongoObjectId = {
   projection: {
@@ -12,36 +26,60 @@ const noMongoObjectId = {
 };
 
 export abstract class EntityRepositoryBase<TModel extends Entity>
-  implements IMongoEntityRepository<TModel> {
-  constructor(database: Db) {
+  implements IMongoEntityRepository<TModel>, IMongoRepository {
+  constructor(context: DbRequestContext) {
     logger.debug(`Entering new ${EntityRepositoryBase.name}()`);
-    if (!database) {
+    if (!context) {
       logger.error(
-        `Argument 'database' must have a value, but got ${JSON.stringify(
-          database,
+        `Argument 'context' must have a value, but got ${JSON.stringify(
+          context,
         )}`,
       );
       throw new Error(`Argument 'database' must have a value!`);
     }
 
-    this.database = database;
+    this.dbContext = context;
   }
 
-  protected database: Db;
-  protected abstract references: IEntityRepositoryReferencePopulator<TModel>[];
-  protected abstract readonly collectionName: string;
+  abstract readonly collectionName: string;
+
+  protected dbContext: DbRequestContext;
+  protected abstract references: IEntityRepositoryReferencePopulator[];
   protected abstract readonly factory: new () => TModel;
-  protected get collection(): Collection {
-    return this.database.collection(this.collectionName);
+
+  protected get db(): Db {
+    return this.dbContext.connection.instance;
   }
 
-  // virtual
-  protected processors: ((entity: TModel) => Promise<void>)[] = [];
+  protected get collection(): Collection {
+    return this.db.collection(this.collectionName);
+  }
+
+  /**
+   * Data processors that operate on the data
+   * before it is returned to the caller as part
+   * of all the find operations.
+   *
+   * @virtual
+   */
+  protected findProcessors: ((entity: TModel) => Promise<void>)[] = [];
+  protected get isExemptFromCollectionPermissions(): boolean {
+    return false;
+  }
+  private readonly _permissionsSvc = new PermissionsService();
+
+  protected fetchCollectionPermissionsAsync(): Promise<CollectionPermission[]> {
+    return Promise.resolve([]);
+  }
 
   async findAsync(
     filter: TModel | { [key: string]: unknown } | FilterQuery<unknown>,
   ): Promise<TModel[]> {
     logger.debug(`Entering ${this.findAsync.name}()`);
+
+    if (!(await this.validateCollectionPermissionsAsync(PermissionAction.read)))
+      return null;
+
     const any = await this.collection
       .find(filter, { ...noMongoObjectId })
       .toArray();
@@ -55,13 +93,17 @@ export abstract class EntityRepositoryBase<TModel extends Entity>
       ]);
     }
 
-    return entities;
+    return (entities || []).filter((ent) =>
+      this.validateEntityPermissions(ent, PermissionAction.read),
+    );
   }
 
   async findOneAsync(
     filter: TModel | { [key: string]: unknown },
+    skipPermissionsChecks = false,
   ): Promise<TModel> {
     logger.debug(`Entering ${this.findOneAsync.name}()`);
+
     const one = await this.collection.findOne(filter, {
       ...noMongoObjectId,
     });
@@ -77,7 +119,11 @@ export abstract class EntityRepositoryBase<TModel extends Entity>
       this.executeDataProcessorsAsync(entity),
     ]);
 
-    return entity;
+    if (skipPermissionsChecks) return entity;
+
+    return this.validateEntityPermissions(entity, PermissionAction.read)
+      ? entity
+      : null;
   }
 
   async findAllAsync(): Promise<TModel[]> {
@@ -95,7 +141,9 @@ export abstract class EntityRepositoryBase<TModel extends Entity>
       ]);
     }
 
-    return entities;
+    return (entities || []).filter((ent) =>
+      this.validateEntityPermissions(ent, PermissionAction.read),
+    );
   }
 
   async findOneByIdAsync(id: string): Promise<TModel> {
@@ -110,7 +158,9 @@ export abstract class EntityRepositoryBase<TModel extends Entity>
         this.executeDataProcessorsAsync(entity),
       ]);
 
-      return entity;
+      return this.validateEntityPermissions(entity, PermissionAction.read)
+        ? entity
+        : null;
     }
 
     logger.debug(`No entity was found with id ${id}`);
@@ -124,6 +174,11 @@ export abstract class EntityRepositoryBase<TModel extends Entity>
       logger.error(`Argument 'entity' must have a value to insert.`);
       throw new Error(`Argument 'entity' must have a value to insert.`);
     }
+
+    if (
+      !(await this.validateCollectionPermissionsAsync(PermissionAction.create))
+    )
+      return null;
 
     entity.id = shortid.generate();
 
@@ -151,6 +206,11 @@ export abstract class EntityRepositoryBase<TModel extends Entity>
       throw new Error(`Argument 'entities' must be an array to insert many.`);
     }
 
+    if (
+      !(await this.validateCollectionPermissionsAsync(PermissionAction.create))
+    )
+      return null;
+
     entities.forEach((e) => (e.id = shortid.generate()));
 
     const result = await this.collection.insertMany(entities);
@@ -168,6 +228,23 @@ export abstract class EntityRepositoryBase<TModel extends Entity>
 
   async deleteOneAsync(id: string): Promise<boolean> {
     logger.debug(`Entering ${this.deleteOneAsync.name}(${id})`);
+
+    if (
+      !(await this.validateCollectionPermissionsAsync(
+        PermissionAction.delete,
+        PermissionAction.read,
+      ))
+    )
+      return false;
+
+    const findResult = await this.findOneByIdAsync(id);
+
+    if (
+      !findResult ||
+      !this.validateEntityPermissions(findResult, PermissionAction.delete)
+    )
+      return false;
+
     const result = await this.collection.deleteOne({ id });
 
     return result.deletedCount === 1;
@@ -178,6 +255,27 @@ export abstract class EntityRepositoryBase<TModel extends Entity>
     entity: TModel | { [key: string]: unknown },
   ): Promise<boolean> {
     logger.debug(`Entering ${this.updateOneAsync.name}(${id}, entity)`);
+
+    if (
+      !this.validateCollectionPermissionsAsync(
+        PermissionAction.read,
+        PermissionAction.update,
+      )
+    )
+      return false;
+
+    const findResult = await this.findOneByIdAsync(id);
+
+    if (
+      !findResult ||
+      !this.validateEntityPermissions(findResult, PermissionAction.update)
+    )
+      return false;
+
+    if (Object.keys(entity).length === 0) {
+      throw new EmptyUpdateEntityError(this.collectionName);
+    }
+
     const result = await this.collection.updateOne(
       { id },
       {
@@ -195,13 +293,24 @@ export abstract class EntityRepositoryBase<TModel extends Entity>
     entity: TModel | { [key: string]: unknown },
   ): Promise<boolean> {
     logger.debug(`Entering ${this.updateManyAsync.name}()`);
-    const result = await this.collection.updateMany(filter, {
-      $set: {
-        ...entity,
-      },
-    });
 
-    return result.modifiedCount > 0;
+    if (
+      !this.validateCollectionPermissionsAsync(
+        PermissionAction.update,
+        PermissionAction.read,
+      )
+    )
+      return false;
+
+    const findResults = await this.findAsync(filter);
+
+    if (!Array.isArray(findResults) || findResults.length === 0) return false;
+
+    await Promise.all(
+      findResults.map((ent) => this.updateOneAsync(ent.id, entity)),
+    );
+
+    return true;
   }
 
   async deleteManyAsync(
@@ -209,9 +318,24 @@ export abstract class EntityRepositoryBase<TModel extends Entity>
   ): Promise<boolean> {
     logger.debug(`Entering ${this.deleteManyAsync.name}(filter)`);
 
-    const result = await this.collection.deleteMany(filter);
+    if (!this.validateCollectionPermissionsAsync(PermissionAction.delete))
+      return false;
 
-    return result.deletedCount > 0;
+    const findResults = await this.findAsync(filter);
+
+    if (!Array.isArray(findResults) || findResults.length === 0) return false;
+
+    await Promise.all(findResults.map((ent) => this.deleteOneAsync(ent.id)));
+
+    return true;
+  }
+
+  async dispose(): Promise<void> {
+    await this.dbContext.connection.dispose();
+  }
+
+  setPermissionWaiver(waiver: IPermissionWaiver): void {
+    this.dbContext.waivePermissions = waiver;
   }
 
   private mapMongoDocumentToEntity(doc: unknown): TModel {
@@ -228,14 +352,72 @@ export abstract class EntityRepositoryBase<TModel extends Entity>
   private async processReferencesAsync(entity: TModel): Promise<void[]> {
     if (Array.isArray(this.references) && this.references.length > 0) {
       return Promise.all(
-        this.references.map((ref) => ref.populateReferenceAsync(entity)),
+        this.references.map((ref) =>
+          this.invokeReferencePopulator(ref, entity),
+        ),
       );
     }
   }
 
   private async executeDataProcessorsAsync(entity: TModel): Promise<void[]> {
-    if (Array.isArray(this.processors) && this.processors.length > 0) {
-      return Promise.all(this.processors.map((proc) => proc(entity)));
+    if (Array.isArray(this.findProcessors) && this.findProcessors.length > 0) {
+      return Promise.all(this.findProcessors.map((proc) => proc(entity)));
     }
+  }
+
+  private invokeReferencePopulator(
+    ref: IEntityRepositoryReferencePopulator,
+    entity: TModel,
+  ): Promise<void> {
+    ref.setPermissionWaiver(this.dbContext.waivePermissions);
+
+    switch (ref.referenceType) {
+      case ReferenceType.oneToOne:
+        return (ref as IEntityRepositorySingleReferencePopulator<
+          TModel
+        >).populateReferenceAsync(entity);
+      case ReferenceType.oneToMany:
+        return (ref as IEntityRepositoryManyReferencesPopulator<
+          TModel
+        >).populateReferencesAsync(entity);
+      default:
+        throw new Error(
+          `Could not determine reference type for cross reference population.`,
+        );
+    }
+  }
+
+  private async validateCollectionPermissionsAsync(
+    ...args: PermissionAction[]
+  ): Promise<boolean> {
+    if (this.isExemptFromCollectionPermissions) return true;
+
+    const collectionPerms = await this.fetchCollectionPermissionsAsync();
+
+    if (!Array.isArray(collectionPerms) || collectionPerms.length === 0)
+      return true;
+
+    const isPermitted = collectionPerms.some((p) =>
+      this._permissionsSvc.isUserPermitted(
+        this.dbContext.user,
+        this.dbContext.waivePermissions,
+        p,
+        args,
+      ),
+    );
+
+    return isPermitted;
+  }
+
+  private validateEntityPermissions(
+    entity: Entity,
+    ...actions: PermissionAction[]
+  ): boolean {
+    return this._permissionsSvc.isUserPermitted(
+      this.dbContext.user,
+      this.dbContext.waivePermissions,
+      entity,
+      actions,
+    );
   }
 }
